+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fixed handling of `pytz` timezones.
+- Fixed some formatter's tokens handling.
 
 
 ## [2.0.2] - 2018-05-29